Api4 - Prevent developer error mixing up the `addValue` and `addWhere` functions
